# app/api/swagger.yml
swagger: "2.0"
info:
  title: "Working Sundays API"
  description: "Job scheduling and store management optimization API"
  version: "1.0.0"
  contact:
    name: "Working Sundays Team"
host: "localhost:5000"
basePath: "/"
schemes:
  - "http"
  - "https"
securityDefinitions:
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
    description: "JWT Authorization header using Bearer scheme. Example: Bearer <token>"
paths:
...
  /api/job/init:
    post:
      tags:
        - "Job Creation"
      summary: "Initialize new job"
      description: "Create a new job for the authenticated user"
      security:
        - Bearer: []
      parameters:
        - in: "body"
          name: "body"
          schema:
            type: "object"
            properties:
              username:
                type: "string"
              descriptor:
                type: "object"
      responses:
        201:
          description: "Job created successfully"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
              data:
                type: "object"
        400:
          description: "Invalid request body"
        403:
          description: "Forbidden"
        500:
          description: "Server error"
  /api/jobs/{username}:
    get:
      tags:
        - "Jobs"
      summary: "List all jobs for user"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "username"
          type: "string"
          required: true
      responses:
        200:
          description: "List of job IDs"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
              data:
                type: "array"
                items:
                  type: "string"
        403:
          description: "Forbidden"
        500:
          description: "Server error"
  /api/job/{username}/{jobid}:
    get:
      tags:
        - "Jobs"
      summary: "Get job information"
      description: "Get comprehensive job info including descriptor, files, and status"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "username"
          type: "string"
          required: true
        - in: "path"
          name: "jobid"
          type: "string"
          required: true
      responses:
        200:
          description: "Job information"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
              data:
                type: "object"
        403:
          description: "Forbidden"
        404:
          description: "Job not found"
        500:
          description: "Server error"
    delete:
      tags:
        - "Jobs"
      summary: "Delete job"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "username"
          type: "string"
          required: true
        - in: "path"
          name: "jobid"
          type: "string"
          required: true
      responses:
        200:
          description: "Job deleted"
        404:
          description: "Job not found"
        500:
          description: "Server error"
  /api/job/{username}/{jobid}/stores:
    post:
      tags:
        - "Job Creation"
      summary: "Save job stores"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "username"
          type: "string"
          required: true
        - in: "path"
          name: "jobid"
          type: "string"
          required: true
        - in: "body"
          name: "body"
          required: true
          schema:
            type: "object"
            properties:
              stores:
                type: "object"
                description: "Mapping of store IDs to store objects"
      responses:
        200:
          description: "Stores saved"
        400:
          description: "Invalid stores payload"
        404:
          description: "Job not found"
  /api/job/{username}/{jobid}/stores-with-radius:
    post:
      tags:
        - "Job Creation"
      summary: "Save stores and compute radius"
      description: "Save stores and compute radius values using calculator"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "username"
          type: "string"
          required: true
        - in: "path"
          name: "jobid"
          type: "string"
          required: true
        - in: "body"
          name: "body"
          required: true
          schema:
            type: "object"
            properties:
              stores:
                type: "object"
              radius_calc:
                type: "string"
      responses:
        200:
          description: "Stores processed"
        400:
          description: "Invalid payload"
        404:
          description: "Job not found"
        500:
          description: "Processing failed"
  /api/job/{username}/{jobid}/constraints:
    post:
      tags:
        - "Job Creation"
      summary: "Save job constraints"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "username"
          type: "string"
          required: true
        - in: "path"
          name: "jobid"
          type: "string"
          required: true
        - in: "body"
          name: "body"
          required: true
          schema:
            type: "object"
            properties:
              constraints:
                type: "object"
      responses:
        200:
          description: "Constraints saved"
        400:
          description: "Invalid payload"
        404:
          description: "Job not found"
  /api/job/{username}/{jobid}/settings:
    post:
      tags:
        - "Job Creation"
      summary: "Save job settings"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "username"
          type: "string"
          required: true
        - in: "path"
          name: "jobid"
          type: "string"
          required: true
        - in: "body"
          name: "body"
          required: true
          schema:
            type: "object"
            properties:
              settings:
                type: "object"
      responses:
        200:
          description: "Settings saved"
        400:
          description: "Invalid payload"
        404:
          description: "Job not found"
  /api/job/{username}/{jobid}/calc:
    post:
      tags:
        - "Job Creation"
      summary: "Save value calculator script"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "username"
          type: "string"
          required: true
        - in: "path"
          name: "jobid"
          type: "string"
          required: true
        - in: "body"
          name: "body"
          required: true
          schema:
            type: "object"
            properties:
              calc:
                type: "string"
      responses:
        200:
          description: "Calculator saved"
        400:
          description: "Invalid payload"
        404:
          description: "Job not found"
  /api/job/{username}/{jobid}/export:
    post:
      tags:
        - "Job Creation"
      summary: "Export job snapshot"
      description: "Export full job descriptor snapshot with all configuration"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "username"
          type: "string"
          required: true
        - in: "path"
          name: "jobid"
          type: "string"
          required: true
      responses:
        200:
          description: "Export payload"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
              data:
                type: "object"
        404:
          description: "Job not found"
        500:
          description: "Server error"
  /api/job/{username}/{jobid}/finish:
    post:
      tags:
        - "Job Creation"
      summary: "Finalize job initialization"
      description: "Validate and finalize job initialization"
      security:
        - Bearer: []
      parameters:
        - in: "path"
          name: "username"
          type: "string"
          required: true
        - in: "path"
          name: "jobid"
          type: "string"
          required: true
      responses:
        200:
          description: "Finalization result"
          schema:
            type: "object"
            properties:
              success:
                type: "boolean"
              data:
                type: "object"
        404:
          description: "Job not found"
        500:
          description: "Server error"
...